Decouple IconEntry and IconPack
